\name{phenotype.snp}



\alias{phenotype.snp}



\docType{data}



\title{
Continuous phenotype dataset
}



\description{
The phenotype data is a numerical vector of length 51. It represents 51
measured phenotypes for 51 laboratory mouse strains. It is to be used as a
dependent variable in combination with the SNP genotype data, in order to
compute the association between each of the SNPs and the phenotype.
}



\usage{data(phenotype.snp)}



\format{
A numerical vector with 51 elements (laboratory mice) which correspond to
the rows of he genotype data.
}



\examples{
data(phenotype.snp)
}




\value{
Vector of 51 metric elements, representing phenotypes measured for 51
laboratory mice.
}



\keyword{dataset}
