\name{phenotype.saap}


\alias{phenotype.saap}


\docType{data}


\title{
Phenotype dataset
}


\description{
The phenotype data is a numerical vector of length 120. It represents 120
measured phenotypes for 120 organisms. We used it as a dependent variable
in combination with the genotype.saap data, and quantified the association
between each of the SAAP and the phenotype.
}

\usage{data(phenotype.saap)}


\format{
A numerical vecotr with 120 elements (organisms) which correspond to the rows
of he genotype data.
}



\examples{
data(phenotype.saap)
}




\value{
Vector of 51 metric elements, representing phenotypes measured for 120
organisms.
}



\keyword{dataset}
